angular/src/app/questionnaire/questions/glazing-type-question/glazing-type-question.component.spec.ts
import {async, ComponentFixture, TestBed} from '@angular/core/testing';

import {ResponseData} from '../../../shared/response-data/response-data';
import {By} from '@angular/platform-browser';
import {MultipleChoiceQuestionComponent} from "../../common-questions/multiple-choice-question/multiple-choice-question.component";
import {GlazingTypeQuestionComponent} from "./glazing-type-question.component";
import {GlazingType} from "./glazing-type";

describe('GlazingTypeQuestionComponent', () => {
    let component: GlazingTypeQuestionComponent;
    let fixture: ComponentFixture<GlazingTypeQuestionComponent>;
    let responseData: ResponseData;

    beforeEach(async(() => {
        TestBed.configureTestingModule({
            declarations: [GlazingTypeQuestionComponent, MultipleChoiceQuestionComponent],
            providers: [ResponseData]
        })
            .compileComponents();
    }));

    beforeEach(() => {
        responseData = TestBed.get(ResponseData);
        fixture = TestBed.createComponent(GlazingTypeQuestionComponent);
        component = fixture.componentInstance;
    });

    it('should create', () => {
        expect(component).toBeTruthy();
    });

    it('should populate with answer in response data', async(() => {
        // given
        responseData.glazingType = GlazingType.Single;

        // when
        fixture.detectChanges();

        fixture.whenStable().then(() => {
            const yesOption = fixture.debugElement.query(By.css('#option-1')).nativeElement;
            expect(yesOption.checked).toBeTruthy();
        });
    }));

    it('should set the response', async(() => {
        fixture.detectChanges();

        fixture.whenStable().then(() => {
            const yesOption = fixture.debugElement.query(By.css('#option-0')).nativeElement;
            yesOption.click();

            expect(responseData.glazingType).toBe(GlazingType.DoNotKnow);
        });
    }));
});
